+++ port_range/__init__.py
@@ -15,17 +15,18 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 import math
+import sys
+from collections import Iterable
 
-try:
-    basestring
-except NameError:  # pragma: no cover
-    basestring = (str, bytes)  # pylint: disable=C0103
 
 try:
     from itertools import imap as iter_map
 except ImportError:  # pragma: no cover
     iter_map = map
 
+PY3 = sys.version_info == 3
+if PY3:
+    basestring = (str, bytes)
 
 __version__ = '1.1.0'
 
@@ -36,7 +37,7 @@ class PortRange(object):
 
     In strict mode (disabled by default) we'll enforce the following rules:
         * port base must be a power of two (offsets not allowed);
-        * port CIDR should not produce overflowing upper bound.
+        * port range must be within the 1-65535 inclusive range.
 
     This mode can be disabled on object creation.
     """
@@ -61,69 +62,62 @@ class PortRange(object):
         self.port_from, self.port_to = self.parse(port_range)
 
     def parse(self, port_range):
-        """ Parse and normalize port range string into a port range. """
-        if isinstance(port_range, basestring) and self.CIDR_SEP in port_range:
-            base, prefix = self.parse_cidr(port_range)
-            port_from, port_to = self._cidr_to_range(base, prefix)
-        else:
-            port_from, port_to = self.parse_range(port_range)
-        if not port_from or not port_to:
-            raise ValueError("Invalid ports.")
-        # Check upper bound
-        if self.strict:
-            # Disallow overflowing upper bound
-            if port_to > self.port_max:
-                raise ValueError("Overflowing upper bound.")
-        else:
-            # Cap upper bound
-            port_to = port_to if port_to < self.port_max else self.port_max
-        return port_from, port_to
-
-    def parse_cidr(self, port_range):
-        """ Split a string and extract port base and CIDR prefix.
-
-        Always returns a list of 2 integers. Defaults to None.
-        """
-        # Separate base and prefix
-        elements = list(iter_map(int, port_range.split(self.CIDR_SEP, 2)))
-        elements += [None, None]
-        base, prefix = elements[:2]
-        # Normalize prefix value
-        if prefix is None:
-            prefix = self.port_lenght
-        # Validates base and prefix values
-        if not base or base < self.port_min or base > self.port_max:
-            raise ValueError("Invalid port base.")
-        if not prefix or prefix < 1 or prefix > self.port_lenght:
-            raise ValueError("Invalid CIDR-like prefix.")
-        # Enable rigorous rules
-        if self.strict:
-            # Disallow offsets
-            if prefix != self.port_lenght and not self._is_power_of_two(base):
-                raise ValueError("Port base is not a power of Two.")
-        return base, prefix
-
-    def parse_range(self, port_range):
-        """ Normalize port range to a sorted list of no more than 2 integers.
-
-        Excludes None values while parsing.
-        """
+        """ Parse and normalize a string or iterable into a port range. """
+        # Any string containing a CIDR separator is parsed as a CIDR-like
+        # notation, others as a range or single port.
+        cidr_notation = False
         if isinstance(port_range, basestring):
-            port_range = port_range.split(self.RANGE_SEP, 2)
-        if not isinstance(port_range, (set, list, tuple)):
+            cidr_notation = self.CIDR_SEP in port_range
+            separator = self.CIDR_SEP if cidr_notation else self.RANGE_SEP
+            port_range = port_range.split(separator, 1)
+
+        # We expect here a list of elements castable to integers.
+        if not isinstance(port_range, Iterable):
             port_range = [port_range]
-        port_range = [int(port)
-                      for port in port_range
-                      if port and int(port)][:2]
-        port_range.sort()
-        # Fill out missing slots by None values
-        port_range += [None] * (2 - len(port_range))
-        port_from, port_to = port_range
-        if not port_from or port_from < self.port_min or \
-                port_from > self.port_max:
-            raise ValueError("Invalid port range lower bound.")
-        if not port_to:
+        try:
+            port_range = list(iter_map(int, port_range))
+        except TypeError:
+            raise ValueError("Can't parse range as a list of integers.")
+
+        # At this point we should have a list of one or two integers.
+        if not 0 < len(port_range) < 3:
+            raise ValueError("Expecting a list of one or two elements.")
+
+        # Transform CIDR notation into a port range and validates it.
+        if cidr_notation:
+            base, prefix = port_range
+            port_range = self._cidr_to_range(base, prefix)
+
+        # Let the parser fix a reverse-ordered range in non-strict mode.
+        if not self.strict:
+            port_range.sort()
+
+        # Get port range bounds.
+        port_from = port_range[0]
+        # Single port gets their upper bound set to None.
+        port_to = port_range[1] if len(port_range) == 2 else None
+
+        # Validate constraints in strict mode.
+        if self.strict:
+            # Disallow out-of-bounds values.
+            if not (self.port_min <= port_from <= self.port_max) or (
+                    port_to is not None and not (
+                        self.port_min <= port_to <= self.port_max)):
+                raise ValueError("Out of bounds.")
+            # Disallow reversed range.
+            if port_to is not None and port_from > port_to:
+                raise ValueError("Invalid reversed port range.")
+
+        # Clamp down lower bound, then cap it.
+        port_from = min([max([port_from, self.port_min]), self.port_max])
+
+        # Single port gets its upper bound aligned to its lower one.
+        if port_to is None:
             port_to = port_from
+
+        # Cap upper bound.
+        port_to = min([port_to, self.port_max])
+
         return port_from, port_to
 
     def __repr__(self):
@@ -177,12 +171,22 @@ class PortRange(object):
         """ Compute a raw upper bound. """
         return base + (2 ** cls._mask(prefix)) - 1
 
-    @classmethod
-    def _cidr_to_range(cls, base, prefix):
+    def _cidr_to_range(self, base, prefix):
         """ Transform a CIDR-like notation into a port range. """
+        # Validates base and prefix values.
+        if not self.port_min <= base <= self.port_max:
+            raise ValueError("Port base out of bounds.")
+        if not 1 <= prefix <= self.port_lenght:
+            raise ValueError("CIDR-like prefix out of bounds.")
+
+        # Disallow offsets in strict mode.
+        if (self.strict and prefix != self.port_lenght
+                and not self._is_power_of_two(base)):
+            raise ValueError("Port base is not a power of two.")
+
         port_from = base
-        port_to = cls._raw_upper_bound(base, prefix)
-        return port_from, port_to
+        port_to = self._raw_upper_bound(base, prefix)
+        return [port_from, port_to]
 
     @property
     def bounds(self):