+++ port_range/tests/test_port_range.py
@@ -26,18 +26,55 @@ class TestPortRange(unittest.TestCase):
         self.assertEqual(PortRange(' 1027 / 15 ').bounds, (1027, 1028))
 
     def test_range_parsing(self):
+        # Normal range.
         self.assertEqual(PortRange('42-4242').bounds, (42, 4242))
+        self.assertEqual(PortRange([42, 4242]).bounds, (42, 4242))
+        self.assertEqual(PortRange(['42', '4242']).bounds, (42, 4242))
+
+        # Single port.
         self.assertEqual(PortRange('42').bounds, (42, 42))
         self.assertEqual(PortRange(42).bounds, (42, 42))
         self.assertEqual(PortRange([42]).bounds, (42, 42))
         self.assertEqual(PortRange(['42']).bounds, (42, 42))
-        self.assertEqual(PortRange([42, 4242]).bounds, (42, 4242))
+
+        # Reversed order.
+        self.assertEqual(PortRange('4242-42').bounds, (42, 4242))
         self.assertEqual(PortRange([4242, 42]).bounds, (42, 4242))
-        self.assertEqual(PortRange([None, 42]).bounds, (42, 42))
-        self.assertEqual(PortRange([42, None]).bounds, (42, 42))
-        self.assertEqual(PortRange([42, None, 32, 3]).bounds, (32, 42))
         self.assertEqual(PortRange((4242, 42)).bounds, (42, 4242))
+        self.assertEqual(PortRange(('4242', '42')).bounds, (42, 4242))
         self.assertEqual(PortRange(set([4242, 42])).bounds, (42, 4242))
+        self.assertEqual(PortRange(set(['4242', '42'])).bounds, (42, 4242))
+
+        # Invalid types and lenght.
+        self.assertRaises(ValueError, PortRange, [None, 42])
+        self.assertRaises(ValueError, PortRange, [42, None])
+        self.assertRaises(ValueError, PortRange, [42, None, 32, 3, -4])
+
+    def test_strict_range_parsing(self):
+        # Normal range.
+        self.assertEqual(PortRange('42-4242', strict=True).bounds, (42, 4242))
+        self.assertEqual(PortRange([42, 4242], strict=True).bounds, (42, 4242))
+        self.assertEqual(PortRange(['42', '4242'], strict=True).bounds, (42, 4242))
+
+        # Single port.
+        self.assertEqual(PortRange('42', strict=True).bounds, (42, 42))
+        self.assertEqual(PortRange(42, strict=True).bounds, (42, 42))
+        self.assertEqual(PortRange([42], strict=True).bounds, (42, 42))
+        self.assertEqual(PortRange(['42'], strict=True).bounds, (42, 42))
+
+        # Reversed order.
+        self.assertRaises(ValueError, PortRange, [4242, 42], True)
+        self.assertRaises(ValueError, PortRange, '4242-42', True)
+        self.assertRaises(ValueError, PortRange, (4242, 42), True)
+        self.assertRaises(ValueError, PortRange, ('4242', '42'), True)
+        # Casting a set to a list might ends up with a naturally sorted list.
+        # self.assertRaises(ValueError, PortRange, set([4242, 42]), True)
+        # self.assertRaises(ValueError, PortRange, set(['4242', '42']), True)
+
+        # Invalid types and lenght.
+        self.assertRaises(ValueError, PortRange, [None, 42], True)
+        self.assertRaises(ValueError, PortRange, [42, None], True)
+        self.assertRaises(ValueError, PortRange, [42, None, 32, 3, -4], True)
 
     def test_cidr_properties(self):
         port = PortRange('1027/15')
@@ -73,7 +110,8 @@ class TestPortRange(unittest.TestCase):
         self.assertEqual(port.port_from, 1234)
         self.assertEqual(port.port_to, 1234)
         self.assertEqual(port.bounds, (1234, 1234))
-        # Upper bound cap
+
+        # Upper-bound cap.
         self.assertEqual(PortRange('64666/3').bounds, (64666, 65535))
 
     def test_output_string(self):
@@ -98,23 +136,49 @@ class TestPortRange(unittest.TestCase):
         self.assertRaises(ValueError, PortRange, None)
         self.assertRaises(ValueError, PortRange, [None])
         self.assertRaises(ValueError, PortRange, [None, None])
+
         # Invalid int
         self.assertRaises(ValueError, PortRange, ' A233 ')
+        self.assertRaises(ValueError, PortRange, '1 2 3 4')
+        self.assertRaises(ValueError, PortRange, 'abcd')
+
         # Test negative values
+        self.assertRaises(ValueError, PortRange, '-24')
         self.assertRaises(ValueError, PortRange, '-24/3')
         self.assertRaises(ValueError, PortRange, '1024/-3')
+        self.assertRaises(ValueError, PortRange, '-1000/-3')
+        self.assertRaises(ValueError, PortRange, '-3/-1000')
+        self.assertRaises(ValueError, PortRange, ' - 24')
+        self.assertRaises(ValueError, PortRange, ' - 24 / 3')
+        self.assertRaises(ValueError, PortRange, ' 1024 / - 3')
+        self.assertRaises(ValueError, PortRange, ' - 1000 / -  3')
+        self.assertRaises(ValueError, PortRange, ' -  3/  - 1000 ')
+
         # Test maximums and minimums
         self.assertRaises(ValueError, PortRange, '1024/0')
         self.assertRaises(ValueError, PortRange, '1024/17')
-        self.assertRaises(ValueError, PortRange, '66666')
-        self.assertRaises(ValueError, PortRange, '0')
         self.assertRaises(ValueError, PortRange, '10-66666', strict=True)
+
+        # Test capping.
+        self.assertEqual(PortRange('0').bounds, (1, 1))
+        self.assertRaises(ValueError, PortRange, '0', True)
+        self.assertEqual(PortRange('66666').bounds, (65535, 65535))
+        self.assertRaises(ValueError, PortRange, '66666', True)
+
+        # Test notation
+        self.assertRaises(ValueError, PortRange, '23/')
+        self.assertRaises(ValueError, PortRange, '/4')
+        self.assertRaises(ValueError, PortRange, '/4')
+        self.assertRaises(ValueError, PortRange, '400-')
+        self.assertRaises(ValueError, PortRange, '123 /  123/ 123')
+        self.assertRaises(ValueError, PortRange, ' 123 - 31123- 1323')
 
     def test_strict_mode(self):
         # Test power of two port base
         PortRange('257', strict=True)
         PortRange('257/16', strict=True)
         self.assertRaises(ValueError, PortRange, '257/4', strict=True)
+
         # Test overflowing upper bound
         self.assertRaises(ValueError, PortRange, '65535/8', strict=True)